+++ pages/create-quiz.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function CreateQuiz() {
+	return <div>create-quize</div>;
+}
