+++ src/app.component/Layout/index.tsx
@@ -5,13 +5,13 @@ import FooterNav from '../navigation/FooterNav';
 interface Props {
 	children: React.ReactNode;
 }
-const notHasNav = [Url.login];
+const hasNav = [Url.home, Url.mypage, Url.updatedQuizzes];
 export default function Layout({ children }: Props) {
 	const router = useRouter();
 	return (
 		<>
 			<main className="px-[20px]">{children}</main>
-			{!notHasNav.includes(router.pathname) && <FooterNav curUrl={router.pathname} />}
+			{hasNav.includes(router.pathname) && <FooterNav curUrl={router.pathname} />}
 		</>
 	);
 }
