+++ src/app.feature/sign/screen/SignScreen.tsx
@@ -1,27 +1,12 @@
-import type { NextPage } from 'next';
-import Router from 'next/router';
-import { useEffect } from 'react';
-import { setCookie } from '@app.modules/cookie/cookie';
+import React from 'react';
 import RunusIcon from '@assets/brand/runus_icon.svg';
 
 import { KAKAO_AUTH_URL, NAVER_AUTH_URL, GOOGLE_AUTH_URL } from '@app.feature/sign/constant/SocialAuthUrl';
 
 import OauthButton from '@app.feature/sign/component/OauthButton';
+import Router from 'next/router';
 
-const Login: NextPage = () => {
-	// const router = useRouter();
-	/* useEffect(() => {
-		const params = new URL(document.location.toString()).searchParams;
-		const isSuccess = params.get('success') === 'success';
-		if (!isSuccess) return; // TO DO : 로그인 에러 페이지로 이동
-		const accessToken = params.get('token') ?? '';
-		const refreshToken = params.get('refresh') ?? '';
-		localStorage.setItem('TEST_TOKEN', accessToken);
-		setCookie('REFRESH_TOKEN', refreshToken, { path: '/', secure: true, sameSite: 'none' });
-		Router.push('/');
-	}, []);
-    */
-
+export default function SignScreen() {
 	return (
 		<div className="w-full  flex flex-col items-center">
 			<div className="mt-[94px]">
@@ -44,7 +29,4 @@ const Login: NextPage = () => {
 			</div>
 		</div>
 	);
-};
-
-export default Login;
-// <div style={{ display: 'none' }}>토큰 GET 용 URL</div>;
+}
